Affichage structure(s) d'un partenaire

+++ api/grants.php
@@ -49,11 +49,38 @@
 	  }
 	}
 
-	// Update a grants in database
+	// Update one grant in database
 	if ($api == 'PUT') {
 	  parse_str(file_get_contents('php://input'), $post_input);
 
-	  $g01 = $tuple->test_input($post_input['membersread']);
+		// Sanitize et affecte les données à des variables locales
+		if(is_array($post_input) && array_key_exists('grant', $post_input)){
+			$grant = $tuple->test_input($post_input['grant']);
+			$actif = $tuple->test_input($post_input['actif']);
+		} else {
+			// Adaptation pour ANGULAR... :-(
+				$dataArray = json_decode(file_get_contents('php://input'),true);
+				if(is_array($dataArray)){
+					$grant = $tuple->test_input($dataArray['grant']);
+					$actif = $tuple->test_input($dataArray['actif']);
+					$id = $tuple->test_input($dataArray['id']);
+				}
+		}
+
+		// Exécution requête. Impossible de supprimer le droit n°1. C'est celui qui est copié
+		// par défaut pour une nouvelle structure.
+	  if ($id != null && $id != 1) {
+	    if ($tuple->updateOne($id, $grant, $actif)) {
+	      echo $tuple->message('Droits modifiés avec succès !',false);
+	    } else {
+	      echo $tuple->message('Erreur lors de la modification des droits !',true);
+	    }
+	  } else {
+	    echo $tuple->message('Droits non trouvés !',true);
+	  }
+
+
+/* 	  $g01 = $tuple->test_input($post_input['membersread']);
 	  $g02 = $tuple->test_input($post_input['memberswrite']);
 	  $g03 = $tuple->test_input($post_input['membersadd']);
 		$g04 = $tuple->test_input($post_input['membersupdate']);
@@ -68,7 +95,7 @@
 		$g13 = $tuple->test_input($post_input['foodsell']);
 		$g14 = $tuple->test_input($post_input['sendnewsletter']);
 
-	  if ($id != null) {
+	  if ($id != null && $id != 1) {
 	    if ($tuple->update($g01, $g02, $g03, $g04, $g05, $g06, $g07, $g08, $g09, $g10, $g11, $g12, $g13, $g14, $id)) {
 	      echo $tuple->message('Droits modifiés avec succès !',false);
 	    } else {
@@ -76,12 +103,12 @@
 	    }
 	  } else {
 	    echo $tuple->message('Droits non trouvés !',true);
-	  }
+	  } */
 	}
 
 	// Delete a grants from database
 	if ($api == 'DELETE') {
-	  if ($id != null) {
+	  if ($id != null && $id != 1) {
 	    if ($tuple->delete($id)) {
 	      echo $tuple->message('Droits supprimés !', false);
 	    } else {
